requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
httptools==0.4.0
pydantic==1.8.2
python-dotenv==0.19.2
python-socketio==5.3.0
google-cloud-speech==2.5.0
google-cloud-translate==2.0.1
firebase-admin==5.0.0